Extract public functions to `ImageCommentsLoader` protocol so that a spy implementation can conforms to it when test-drive UI
The price we pay is that we lose information of returned error type again. This is acceptable since in the future we may have another
concrete type conforms to `ImageCommentsLoader ` (e.g. a `LocalImageCommentLoader`) and returns its own error type. It's then the
responsibility of the client the check the error type when needed.